requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-dateutil>=2.8.2
coreapi>=2.3.3
pyyaml>=6.0.2